improves database and user model

+++ app/models/connected_database.rb
@@ -3,13 +3,18 @@ class ConnectedDatabase < ActiveRecord::Base
   
   def self.create_database(options={})
     unless options[:name].blank?
-      connection.execute "CREATE DATABASE `#{options[:name]}`;"
+      connection.execute "CREATE DATABASE IF NOT EXISTS `#{options[:name]}`;"
     end
   end
   
   def self.create_user(options={})
     unless options[:name].blank? && options[:password].blank?
-      connection.execute "CREATE USER '#{options[:name]}'@'localhost' IDENTIFIED BY '#{options[:password]}';"
+      res = connection.execute("SELECT COUNT(*) FROM user WHERE User = '#{options[:name]}';")
+      row = res.fetch_row
+      puts "xxxx" + row[0].to_s
+      if row[0].to_i < 1
+        connection.execute "CREATE USER '#{options[:name]}'@'localhost' IDENTIFIED BY '#{options[:password]}';"
+      end
     end
   end
   
@@ -21,26 +26,41 @@ class ConnectedDatabase < ActiveRecord::Base
   
   def self.destroy_database(options={})
     unless options[:name].blank?
-      connection.execute "DROP DATABASE `#{options[:name]}`;"
-    end
-  end
-  
-  def self.destroy_user(options={})
-    unless options[:name].blank?
-      connection.execute "DROP USER '#{options[:name]}'@'localhost';"
+      connection.execute "DROP DATABASE IF EXISTS `#{options[:name]}`;"
     end
   end
   
   def self.change_user_password(options={})
     unless options[:name].blank? && options[:password].blank?
-      connection.execute "SET PASSWORD FOR '#{options[:name]}'@'localhost' = PASSWORD( '#{options[:password]}' );"
+      res = connection.execute("SELECT COUNT(*) FROM user WHERE User = '#{options[:name]}';")
+      row = res.fetch_row
+
+      if row[0].to_i > 0
+        connection.execute "SET PASSWORD FOR '#{options[:name]}'@'localhost' = PASSWORD( '#{options[:password]}' );"
+      end
     end
   end
   
   def self.rename_user(options={})
     unless options[:name].blank? && options[:oldname].blank?
+      res = connection.execute("SELECT COUNT(*) FROM user WHERE User = '#{options[:oldname]}';")
+      row = res.fetch_row
+      
+      if row[0].to_i > 0
+        connection.execute "RENAME USER '#{options[:oldname]}'@'localhost' TO '#{options[:name]}'@'localhost';"
+        connection.execute "FLUSH PRIVILEGES;"
+      end
+    end
+  end
+  
+  def self.destroy_user(options={})
+    unless options[:name].blank?
       res = connection.execute("SELECT COUNT(*) FROM user WHERE User = '#{options[:name]}';")
-      connection.execute "RENAME USER '#{options[:oldname]}'@'localhost' TO '#{options[:name]}'@'localhost';" if res.fetch_row.to_i != 0
+      row = res.fetch_row
+
+      if row[0].to_i > 0
+        connection.execute "DROP USER '#{options[:name]}'@'localhost';"
+      end
     end
   end
 end
